feat(database): add user migrations and unify flyway tooling
Add Flyway SQL and CSV seed files for User, UserSettings, and UserInformation, plus user-related stored procedures and schema updates.

Align database and PowerShell tooling scripts to the current repository layout, update GetRuleByName to the ATAPUtilities schema, and remove the obsolete Invoke-DatabaseFlywayMigrations entrypoint.

+++ Database/Flyway/SQL/V00.01.000010__Create_ATAPUtilities_Core_Schema.sql
@@ -313,7 +313,7 @@ CREATE TABLE ATAPUtilities.UserInformation (
     Email                VARBINARY(MAX)       NULL,
     Phone                VARBINARY(MAX)       NULL,
     Role                 VARBINARY(MAX)       NULL,
-    EncryptionKeyVersion INT              NOT NULL CONSTRAINT DF_UserInformation_EncryptionKeyVersion DEFAULT 1,
+    EncryptionKeyVersion TINYINT              NOT NULL CONSTRAINT DF_UserInformation_EncryptionKeyVersion DEFAULT 1,
     CONSTRAINT PK_UserInformation          PRIMARY KEY CLUSTERED (UserId),
     CONSTRAINT FK_UserInformation_User     FOREIGN KEY (UserId) REFERENCES ATAPUtilities.[User] (UserId),
     CONSTRAINT CK_UserInformation_Contact  CHECK (
